requirements.txt
alembic==1.8.1
click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
